requirements.txt
Flask==2.3.3
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0
Werkzeug==2.3.7 